+++ Source/ACE.Server/WorldObjects/Creature_Equipment.cs
@@ -56,8 +56,7 @@ namespace ACE.Server.WorldObjects
             // todo see if the wielded location is in use, if so, return false
 
             worldObject.SetProperty(PropertyInt.CurrentWieldedLocation, wieldedLocation);
-            worldObject.SetProperty(PropertyInstanceId.Wielder, (int)Biota.Id);
-
+            worldObject.SetProperty(PropertyInstanceId.Wielder, Biota.Id);
             EquippedObjects[worldObject.Guid] = worldObject;
 
             return true;
